requirements.txt
amqp==5.0.6
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
decorator==5.0.9
Django==3.2.5
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==21.1
django-shell-plus==1.1.6
django-silk==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
fake-useragent==0.1.11
flake8==3.9.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==4.0.0
gprof2dot==2021.2.21
h11==0.12.0
idna==3.2
inflection==0.5.1
ipython==7.25.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
outcome==1.1.0
packaging==21.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wsproto==1.0.0
